tty_interface: Exit when backspace is pressed on empty input
This is how the : prompt in vim handles backspace on empty input, so it
feels convenient.

This patch may be bad for people who press backspace repeatedly to clear
the input line and accidentally press backspace once too often, exiting
fzy; I use ^U to clear the input line, so it's not a problem for me.